Consultas a 2 o más tablas && Alias

+++ clase2/crear-tabla-proveedores.sql
@@ -2,7 +2,7 @@
 
 create table proveedores
 (
-    idProveedor int auto_increment primary key,
+    idProveedor tinyint unsigned auto_increment primary key,
     razonsocial varchar(100) unique not null,
     cuit int not null,
     telefono int not null,
